Completed implementation of iterator, started List
This commit is bigger than it should be...

I started on the List, then realized it was easier to start
on the iterator, so I did that instead. The iterator was
fairly trivial, and seems to compile fine, although there's
no way to know if it's full of bugs or not until the List is
complete and can be tested.