Memory Management working (no leaks)
Very basic recursive memory management is working, with valgrind
reporting no leaks, but some odd problems.

Did a few re-organizations/refactorings to facilitate mem managment,
such as taking allocated nodes from the end of the allocated block
and decrementing.

There is still a lot to do in terms of memory management, such
as playing with the size of blocks allocated, allocating in
exponentially increasing blocks, figuring out if 16 is really a good
default size, etc. etc.

Right now everything is completely arbitrary and really based on
nothing, but it does apparently work.