Added insertion sort as last pass of quicksort
From my very non-scientific testing, there is some small gain in
performance obtained by doing this. Sorting (1 << 24 = ~16,000,000)
integers goes from around 20 seconds to around 19, which is a 5%
improvement, which is certainly not nothing. Based on that I think it's
a worthwhile change.

The changes in main.cpp are mostly to add a test to make sure that the
sorted list ends up also being the right length (e.g. we're not
accidentally dropping elements in the sorting procedure).