Mergesort implemented and working
Recursive mergesort with a basecase insertion sort for small lists is
working and seriously outperforming my quicksort :(. It's just under
twice as fast for a list of ~16,000,000 integers.

Additionally, it lends itself to being generalized for an arbitrary
boolean comparison function far better than my quicksort, because it
doesn't need to screw around with equality checks.

TODO - template .sort() for arbitrary comparator functors